消除x64dmmain的警告，让qqdemo在x64下正常运行

+++ DmMain/inc/Common/ActiveX/DMActiveXHelper.h
@@ -412,7 +412,7 @@ namespace DM
 		// Invoke a method by DISPID with N parameters
 		HRESULT InvokeN(DISPID dispid, VARIANT* pvarParams, int nParams, VARIANT* pvarRet = NULL) throw()
 		{
-			DISPPARAMS dispparams = { pvarParams, NULL, nParams, 0};
+			DISPPARAMS dispparams = { pvarParams, NULL, (UINT)nParams, 0};
 			return p->Invoke(dispid, IID_NULL, LOCALE_USER_DEFAULT, DISPATCH_METHOD, &dispparams, pvarRet, NULL, NULL);
 		}
 		// Invoke a method by name with Nparameters
